Create schema-layer for Zod schemas
