+++ packages/ensnode-sdk/src/ensindexer/indexing-status/deserialize/cross-chain-indexing-status-snapshot.ts
@@ -1,16 +1,12 @@
-import z, { prettifyError } from "zod/v4";
+import { prettifyError, z } from "zod/v4";
 
-import { makeUnixTimestampSchema } from "../../../shared/zod-schemas";
+import type { CrossChainIndexingStatusSnapshot } from "../cross-chain-indexing-status-snapshot";
 import {
-  type CrossChainIndexingStatusSnapshot,
-  CrossChainIndexingStrategyIds,
-} from "../cross-chain-indexing-status-snapshot";
+  buildUnvalidatedCrossChainIndexingStatusSnapshot,
+  makeCrossChainIndexingStatusSnapshotSchema,
+  makeSerializedCrossChainIndexingStatusSnapshotSchema,
+} from "../schema/cross-chain-indexing-status-snapshot";
 import type { SerializedCrossChainIndexingStatusSnapshot } from "../serialize/cross-chain-indexing-status-snapshot";
-import { makeCrossChainIndexingStatusSnapshotSchema } from "../validate/cross-chain-indexing-status-snapshot";
-import {
-  buildUnvalidatedOmnichainIndexingStatusSnapshot,
-  makeSerializedOmnichainIndexingStatusSnapshotSchema,
-} from "./omnichain-indexing-status-snapshot";
 
 /**
  * Deserialize an {@link CrossChainIndexingStatusSnapshot} object.
@@ -35,36 +31,3 @@ export function deserializeCrossChainIndexingStatusSnapshot(
 
   return parsed.data;
 }
-
-/**
- * Build unvalidated cross-chain indexing status snapshot to be validated.
- *
- * Return type is intentionally "unknown" to enforce validation by
- * {@link makeCrossChainIndexingStatusSnapshotSchema} call.
- */
-export function buildUnvalidatedCrossChainIndexingStatusSnapshot(
-  serializedCrossChainIndexingStatusSnapshot: SerializedCrossChainIndexingStatusSnapshot,
-): unknown {
-  const { strategy, slowestChainIndexingCursor, snapshotTime, omnichainSnapshot } =
-    serializedCrossChainIndexingStatusSnapshot;
-
-  return {
-    strategy,
-    slowestChainIndexingCursor,
-    snapshotTime,
-    omnichainSnapshot: buildUnvalidatedOmnichainIndexingStatusSnapshot(omnichainSnapshot),
-  };
-}
-
-/**
- * Makes Zod schema for {@link SerializedCrossChainIndexingStatusSnapshot}
- */
-export const makeSerializedCrossChainIndexingStatusSnapshotSchema = (
-  valueLabel: string = "Cross-chain Indexing Status Snapshot",
-) =>
-  z.object({
-    strategy: z.enum(CrossChainIndexingStrategyIds),
-    slowestChainIndexingCursor: makeUnixTimestampSchema(valueLabel),
-    snapshotTime: makeUnixTimestampSchema(valueLabel),
-    omnichainSnapshot: makeSerializedOmnichainIndexingStatusSnapshotSchema(valueLabel),
-  });
