+++ packages/ensnode-sdk/src/ensindexer/indexing-status/validate/omnichain-indexing-status-snapshot.ts
@@ -1,96 +1,7 @@
-import z, { prettifyError } from "zod/v4";
-import type { ParsePayload } from "zod/v4/core";
+import { prettifyError } from "zod/v4";
 
-import type { ChainId } from "../../../shared/types";
-import { makeChainIdSchema, makeUnixTimestampSchema } from "../../../shared/zod-schemas";
-import {
-  ChainIndexingStatusIds,
-  type ChainIndexingStatusSnapshot,
-  type ChainIndexingStatusSnapshotCompleted,
-  type ChainIndexingStatusSnapshotQueued,
-} from "../chain-indexing-status-snapshot";
-import {
-  type ChainIndexingStatusSnapshotForOmnichainIndexingStatusSnapshotBackfill,
-  getOmnichainIndexingStatus,
-  OmnichainIndexingStatusIds,
-  type OmnichainIndexingStatusSnapshot,
-  type OmnichainIndexingStatusSnapshotFollowing,
-} from "../omnichain-indexing-status-snapshot";
-import { makeChainIndexingStatusSnapshotSchema } from "./chain-indexing-status-snapshot";
-
-/**
- * Check if Chain Indexing Status Snapshots fit the 'unstarted' overall status
- * snapshot requirements:
- * - All chains are guaranteed to have a status of "queued".
- *
- * Note: This function narrows the {@link ChainIndexingStatusSnapshot} type to
- * {@link ChainIndexingStatusSnapshotQueued}.
- */
-export function checkChainIndexingStatusSnapshotsForOmnichainStatusSnapshotUnstarted(
-  chains: ChainIndexingStatusSnapshot[],
-): chains is ChainIndexingStatusSnapshotQueued[] {
-  return chains.every((chain) => chain.chainStatus === ChainIndexingStatusIds.Queued);
-}
-
-/**
- * Check if Chain Indexing Status Snapshots fit the 'backfill' overall status
- * snapshot requirements:
- * - At least one chain is guaranteed to be in the "backfill" status.
- * - Each chain is guaranteed to have a status of either "queued",
- *   "backfill" or "completed".
- *
- * Note: This function narrows the {@link ChainIndexingStatusSnapshot} type to
- * {@link ChainIndexingStatusSnapshotForOmnichainIndexingStatusSnapshotBackfill}.
- */
-export function checkChainIndexingStatusSnapshotsForOmnichainStatusSnapshotBackfill(
-  chains: ChainIndexingStatusSnapshot[],
-): chains is ChainIndexingStatusSnapshotForOmnichainIndexingStatusSnapshotBackfill[] {
-  const atLeastOneChainInTargetStatus = chains.some(
-    (chain) => chain.chainStatus === ChainIndexingStatusIds.Backfill,
-  );
-  const otherChainsHaveValidStatuses = chains.every(
-    (chain) =>
-      chain.chainStatus === ChainIndexingStatusIds.Queued ||
-      chain.chainStatus === ChainIndexingStatusIds.Backfill ||
-      chain.chainStatus === ChainIndexingStatusIds.Completed,
-  );
-
-  return atLeastOneChainInTargetStatus && otherChainsHaveValidStatuses;
-}
-
-/**
- * Checks if Chain Indexing Status Snapshots fit the 'completed' overall status
- * snapshot requirements:
- * - All chains are guaranteed to have a status of "completed".
- *
- * Note: This function narrows the {@link ChainIndexingStatusSnapshot} type to
- * {@link ChainIndexingStatusSnapshotCompleted}.
- */
-export function checkChainIndexingStatusSnapshotsForOmnichainStatusSnapshotCompleted(
-  chains: ChainIndexingStatusSnapshot[],
-): chains is ChainIndexingStatusSnapshotCompleted[] {
-  const allChainsHaveValidStatuses = chains.every(
-    (chain) => chain.chainStatus === ChainIndexingStatusIds.Completed,
-  );
-
-  return allChainsHaveValidStatuses;
-}
-
-/**
- * Checks Chain Indexing Status Snapshots fit the 'following' overall status
- * snapshot requirements:
- * - At least one chain is guaranteed to be in the "following" status.
- * - Any other chain can have any status.
- */
-export function checkChainIndexingStatusSnapshotsForOmnichainStatusSnapshotFollowing(
-  chains: ChainIndexingStatusSnapshot[],
-): chains is ChainIndexingStatusSnapshot[] {
-  const allChainsHaveValidStatuses = chains.some(
-    (chain) => chain.chainStatus === ChainIndexingStatusIds.Following,
-  );
-
-  return allChainsHaveValidStatuses;
-}
+import type { OmnichainIndexingStatusSnapshot } from "../omnichain-indexing-status-snapshot";
+import { makeOmnichainIndexingStatusSnapshotSchema } from "../schema/omnichian-indexing-status-snapshot";
 
 /**
  * Validate an {@link OmnichainIndexingStatusSnapshot} object.
@@ -107,307 +18,3 @@ export function validateOmnichainIndexingStatusSnapshot(
 
   return parsed.data;
 }
-
-/**
- * Invariant: For omnichain snapshot,
- * `omnichainStatus` is set based on the snapshots of individual chains.
- */
-export function invariant_omnichainSnapshotStatusIsConsistentWithChainSnapshot(
-  ctx: ParsePayload<OmnichainIndexingStatusSnapshot>,
-) {
-  const snapshot = ctx.value;
-  const chains = Array.from(snapshot.chains.values());
-  const expectedOmnichainStatus = getOmnichainIndexingStatus(chains);
-  const actualOmnichainStatus = snapshot.omnichainStatus;
-
-  if (expectedOmnichainStatus !== actualOmnichainStatus) {
-    ctx.issues.push({
-      code: "custom",
-      input: snapshot,
-      message: `'${actualOmnichainStatus}' is an invalid omnichainStatus. Expected '${expectedOmnichainStatus}' based on the statuses of individual chains.`,
-    });
-  }
-}
-
-/**
- * Invariant: For omnichain status snapshot,
- * `omnichainIndexingCursor` is lower than the earliest start block
- * across all queued chains.
- *
- * Note: if there are no queued chains, the invariant holds.
- */
-export function invariant_omnichainIndexingCursorLowerThanEarliestStartBlockAcrossQueuedChains(
-  ctx: ParsePayload<OmnichainIndexingStatusSnapshot>,
-) {
-  const snapshot = ctx.value;
-  const queuedChains = Array.from(snapshot.chains.values()).filter(
-    (chain) => chain.chainStatus === ChainIndexingStatusIds.Queued,
-  );
-
-  // there are no queued chains
-  if (queuedChains.length === 0) {
-    // the invariant holds
-    return;
-  }
-
-  const queuedChainStartBlocks = queuedChains.map((chain) => chain.config.startBlock.timestamp);
-  const queuedChainEarliestStartBlock = Math.min(...queuedChainStartBlocks);
-
-  // there are queued chains
-  // the invariant holds if the omnichain indexing cursor is lower than
-  // the earliest start block across all queued chains
-  if (snapshot.omnichainIndexingCursor >= queuedChainEarliestStartBlock) {
-    ctx.issues.push({
-      code: "custom",
-      input: snapshot,
-      message:
-        "`omnichainIndexingCursor` must be lower than the earliest start block across all queued chains.",
-    });
-  }
-}
-
-/**
- * Invariant: For omnichain status snapshot,
- * `omnichainIndexingCursor` is lower than or equal to
- * the highest `backfillEndBlock` across all backfill chains.
- *
- * Note: if there are no backfill chains, the invariant holds.
- */
-export function invariant_omnichainIndexingCursorLowerThanOrEqualToLatestBackfillEndBlockAcrossBackfillChains(
-  ctx: ParsePayload<OmnichainIndexingStatusSnapshot>,
-) {
-  const snapshot = ctx.value;
-  const backfillChains = Array.from(snapshot.chains.values()).filter(
-    (chain) => chain.chainStatus === ChainIndexingStatusIds.Backfill,
-  );
-
-  // there are no backfill chains
-  if (backfillChains.length === 0) {
-    // the invariant holds
-    return;
-  }
-
-  const backfillEndBlocks = backfillChains.map((chain) => chain.backfillEndBlock.timestamp);
-  const highestBackfillEndBlock = Math.max(...backfillEndBlocks);
-
-  // there are backfill chains
-  // the invariant holds if the omnichainIndexingCursor is lower than or
-  // equal to the highest backfillEndBlock across all backfill chains.
-  if (snapshot.omnichainIndexingCursor > highestBackfillEndBlock) {
-    ctx.issues.push({
-      code: "custom",
-      input: snapshot,
-      message:
-        "`omnichainIndexingCursor` must be lower than or equal to the highest `backfillEndBlock` across all backfill chains.",
-    });
-  }
-}
-
-/**
- * Invariant: For omnichain status snapshot,
- * `omnichainIndexingCursor` is same as the highest latestIndexedBlock
- * across all indexed chains.
- *
- * Note: if there are no indexed chains, the invariant holds.
- */
-export function invariant_omnichainIndexingCursorIsEqualToHighestLatestIndexedBlockAcrossIndexedChain(
-  ctx: ParsePayload<OmnichainIndexingStatusSnapshot>,
-) {
-  const snapshot = ctx.value;
-  const indexedChains = Array.from(snapshot.chains.values()).filter(
-    (chain) =>
-      chain.chainStatus === ChainIndexingStatusIds.Backfill ||
-      chain.chainStatus === ChainIndexingStatusIds.Completed ||
-      chain.chainStatus === ChainIndexingStatusIds.Following,
-  );
-
-  // there are no indexed chains
-  if (indexedChains.length === 0) {
-    // the invariant holds
-    return;
-  }
-
-  const indexedChainLatestIndexedBlocks = indexedChains.map(
-    (chain) => chain.latestIndexedBlock.timestamp,
-  );
-  const indexedChainHighestLatestIndexedBlock = Math.max(...indexedChainLatestIndexedBlocks);
-
-  // there are indexed chains
-  // the invariant holds if the omnichain indexing cursor is same as
-  // the highest latestIndexedBlock across all indexed chains
-  if (snapshot.omnichainIndexingCursor !== indexedChainHighestLatestIndexedBlock) {
-    ctx.issues.push({
-      code: "custom",
-      input: snapshot,
-      message:
-        "`omnichainIndexingCursor` must be same as the highest `latestIndexedBlock` across all indexed chains.",
-    });
-  }
-}
-
-/**
- * Invariant: For omnichain status snapshot 'unstarted',
- * all chains must have "queued" status.
- */
-export function invariant_omnichainSnapshotUnstartedHasValidChains(
-  ctx: ParsePayload<Map<ChainId, ChainIndexingStatusSnapshot>>,
-) {
-  const chains = ctx.value;
-  const hasValidChains = checkChainIndexingStatusSnapshotsForOmnichainStatusSnapshotUnstarted(
-    Array.from(chains.values()),
-  );
-
-  if (hasValidChains === false) {
-    ctx.issues.push({
-      code: "custom",
-      input: chains,
-      message: `For omnichain status snapshot 'unstarted', all chains must have "queued" status.`,
-    });
-  }
-}
-
-/**
- * Invariant: For omnichain status snapshot 'backfill',
- * at least one chain must be in "backfill" status and
- * each chain has to have a status of either "queued", "backfill"
- * or "completed".
- */
-export function invariant_omnichainStatusSnapshotBackfillHasValidChains(
-  ctx: ParsePayload<Map<ChainId, ChainIndexingStatusSnapshot>>,
-) {
-  const chains = ctx.value;
-  const hasValidChains = checkChainIndexingStatusSnapshotsForOmnichainStatusSnapshotBackfill(
-    Array.from(chains.values()),
-  );
-
-  if (hasValidChains === false) {
-    ctx.issues.push({
-      code: "custom",
-      input: chains,
-      message: `For omnichain status snapshot 'backfill', at least one chain must be in "backfill" status and each chain has to have a status of either "queued", "backfill" or "completed".`,
-    });
-  }
-}
-
-/**
- * Invariant: For omnichain status snapshot 'completed',
- * all chains must have "completed" status.
- */
-export function invariant_omnichainStatusSnapshotCompletedHasValidChains(
-  ctx: ParsePayload<Map<ChainId, ChainIndexingStatusSnapshot>>,
-) {
-  const chains = ctx.value;
-  const hasValidChains = checkChainIndexingStatusSnapshotsForOmnichainStatusSnapshotCompleted(
-    Array.from(chains.values()),
-  );
-
-  if (hasValidChains === false) {
-    ctx.issues.push({
-      code: "custom",
-      input: chains,
-      message: `For omnichain status snapshot 'completed', all chains must have "completed" status.`,
-    });
-  }
-}
-
-/**
- * Invariant: For omnichain status snapshot 'following',
- * at least one chain must be in 'following' status.
- */
-export function invariant_omnichainStatusSnapshotFollowingHasValidChains(
-  ctx: ParsePayload<OmnichainIndexingStatusSnapshotFollowing>,
-) {
-  const snapshot = ctx.value;
-  const hasValidChains = checkChainIndexingStatusSnapshotsForOmnichainStatusSnapshotFollowing(
-    Array.from(snapshot.chains.values()),
-  );
-
-  if (hasValidChains === false) {
-    ctx.issues.push({
-      code: "custom",
-      input: snapshot,
-      message: "For omnichainStatus 'following', at least one chain must be in 'following' status.",
-    });
-  }
-}
-
-export const makeChainIndexingStatusSnapshotsSchema = (valueLabel: string = "Value") =>
-  z.map(makeChainIdSchema(), makeChainIndexingStatusSnapshotSchema(valueLabel), {
-    error:
-      "Chains indexing statuses must be an object mapping valid chain IDs to their indexing status snapshots.",
-  });
-
-/**
- * Makes Zod schema for {@link OmnichainIndexingStatusSnapshotUnstarted}
- */
-const makeOmnichainIndexingStatusSnapshotUnstartedSchema = (valueLabel?: string) =>
-  z.strictObject({
-    omnichainStatus: z.literal(OmnichainIndexingStatusIds.Unstarted),
-    chains: makeChainIndexingStatusSnapshotsSchema(valueLabel)
-      .check(invariant_omnichainSnapshotUnstartedHasValidChains)
-      .transform((chains) => chains as Map<ChainId, ChainIndexingStatusSnapshotQueued>),
-    omnichainIndexingCursor: makeUnixTimestampSchema(valueLabel),
-  });
-
-/**
- * Makes Zod schema for {@link OmnichainIndexingStatusSnapshotBackfill}
- */
-const makeOmnichainIndexingStatusSnapshotBackfillSchema = (valueLabel?: string) =>
-  z.strictObject({
-    omnichainStatus: z.literal(OmnichainIndexingStatusIds.Backfill),
-    chains: makeChainIndexingStatusSnapshotsSchema(valueLabel)
-      .check(invariant_omnichainStatusSnapshotBackfillHasValidChains)
-      .transform(
-        (chains) =>
-          chains as Map<
-            ChainId,
-            ChainIndexingStatusSnapshotForOmnichainIndexingStatusSnapshotBackfill
-          >,
-      ),
-    omnichainIndexingCursor: makeUnixTimestampSchema(valueLabel),
-  });
-
-/**
- * Makes Zod schema for {@link OmnichainIndexingStatusSnapshotCompleted}
- */
-const makeOmnichainIndexingStatusSnapshotCompletedSchema = (valueLabel?: string) =>
-  z.strictObject({
-    omnichainStatus: z.literal(OmnichainIndexingStatusIds.Completed),
-    chains: makeChainIndexingStatusSnapshotsSchema(valueLabel)
-      .check(invariant_omnichainStatusSnapshotCompletedHasValidChains)
-      .transform((chains) => chains as Map<ChainId, ChainIndexingStatusSnapshotCompleted>),
-    omnichainIndexingCursor: makeUnixTimestampSchema(valueLabel),
-  });
-
-/**
- * Makes Zod schema for {@link OmnichainIndexingStatusSnapshotFollowing}
- */
-const makeOmnichainIndexingStatusSnapshotFollowingSchema = (valueLabel?: string) =>
-  z.strictObject({
-    omnichainStatus: z.literal(OmnichainIndexingStatusIds.Following),
-    chains: makeChainIndexingStatusSnapshotsSchema(valueLabel),
-    omnichainIndexingCursor: makeUnixTimestampSchema(valueLabel),
-  });
-
-/**
- * Omnichain Indexing Snapshot Schema
- *
- * Makes a Zod schema definition for validating indexing snapshot
- * across all chains indexed by ENSIndexer instance.
- */
-export const makeOmnichainIndexingStatusSnapshotSchema = (
-  valueLabel: string = "Omnichain Indexing Snapshot",
-) =>
-  z
-    .discriminatedUnion("omnichainStatus", [
-      makeOmnichainIndexingStatusSnapshotUnstartedSchema(valueLabel),
-      makeOmnichainIndexingStatusSnapshotBackfillSchema(valueLabel),
-      makeOmnichainIndexingStatusSnapshotCompletedSchema(valueLabel),
-      makeOmnichainIndexingStatusSnapshotFollowingSchema(valueLabel),
-    ])
-    .check(invariant_omnichainSnapshotStatusIsConsistentWithChainSnapshot)
-    .check(invariant_omnichainIndexingCursorLowerThanEarliestStartBlockAcrossQueuedChains)
-    .check(
-      invariant_omnichainIndexingCursorLowerThanOrEqualToLatestBackfillEndBlockAcrossBackfillChains,
-    )
-    .check(invariant_omnichainIndexingCursorIsEqualToHighestLatestIndexedBlockAcrossIndexedChain);
